refactor(homepage): rearrange lib folder by service repo layer
