+++ homepage/src/CMS/preview/CardPreview.jsx
@@ -1,5 +1,5 @@
 import Card from '../../components/cards/card';
-import { processCard } from '../../lib/processCard';
+import { processCard } from '../../lib/utils/processCard';
 
 const CardPreview = ({ entry, getAsset, assetsByLocale, locale }) => {
   const assets = assetsByLocale[locale];
