+++ homepage/src/lib/repository/fetchFooter.js
