+++ homepage/src/lib/utils/getJobCardColor.js
