+++ homepage/src/lib/utils/processCard.js
