+++ homepage/src/pages/cards.jsx
@@ -1,7 +1,7 @@
 import Head from 'next/head';
 import contentMapper from '../layouts/contentMapper';
-import { getLayout } from '../lib/getLayout';
-import { getPage } from '../lib/getPage';
+import { getLayout } from '../lib/service/getLayout';
+import { getPage } from '../lib/service/getPage';
 
 /**
  *
